+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.34.1] - 2023-07-31
+
 ### Changes
 
 -   Delaying the initial `getDynamicLoginMethods` until the first feature/routing component renders